# swagger.yaml
swagger: '2.0'
info:
  description: |+
    REST API that provides a requirements classifier. The classifier is based on a Naive Bayes implemented using Apache Mahout (https://mahout.apache.org/)

    This service implements two different classifiers:

    **· Binary classifier:** given a property value V for the *requirement_type* property, the binary classifier builds a classifier to evaluate if a requirement is a member of the property value class (*requirement_type = 'V'*) or if it is not (*requirement_type = 'Prose'*).

    **· Multilabel classifier:** given the domain of property values {V}, the multilabel classifier is used for properties that can have more than one value for each item.

  title: Requirement Classifier API
host: 'https://api.openreq.eu/requirements-classifier'
basePath: /
...
paths:
  /upc/classifier-component/classify:
    post:
      tags:
      - binary-classification-controller
      summary: Classify a list of requirements
      description: 'Given a list of requirements, and using the model stored for the requested company, the requirements are classified  and a recommended label is returned for each requirement (with a level of confidence)'
      operationId: classifyUsingPOST
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: company
        in: query
        description: Proprietary company of the model
        required: true
        type: string
        allowEmptyValue: false
      - name: context
        in: query
        description: 'Whether to apply contextual information analysis or not. If positive, contextual information (i.e. requirements'' position order and hierarchical structure) is used to apply the same tag during the classification process to all members belonging to a same document list structure.'
        required: false
        type: boolean
        default: false
        allowEmptyValue: false
        x-example: false
      - name: property
        in: query
        description: Property of the classifier (i.e. property value of the *requirement_type* field)
        required: true
        type: string
        allowEmptyValue: false
      - in: body
        name: request
        description: Request with the requirements to classify
        required: true
        schema:
          $ref: '#/definitions/Classify requirements list'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/RecommendationList'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /upc/classifier-component/model:
    post:
      tags:
      - binary-classification-controller
      summary: Create a model
      description: "Given a list of requirements, and a specific company and property, a new model is generated and stored in the database. Each model is identified by a company and a property. Therefore, there is only one model per property and company.\n\nThis method is executed in an asynchronous way. As a response you will get an object with a single attribute id (check swagger response below) when the training has started. Once the execution is finished, you will get a response to the endpoint set at the *url* parameter. The format of the response is as follows: \n\n{\n\t\"message\":\"Response message\",\n\t\"id\": \"1562315038067_409\",\n\t\"code\": 200\n}\n\nThe 'id' field can be used to match the synchronous response of the request with the asynchronous response of the response. The 'code' field states the HTTP code of the request."
      operationId: trainUsingPOST
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - name: company
        in: query
        description: Proprietary company of the model
        required: true
        type: string
        allowEmptyValue: false
        x-example: UPC
      - name: property
        in: query
        description: Property of the classifier (i.e. property value of the *requirement_type* field)
        required: true
        type: string
        allowEmptyValue: false
        x-example: requirement
      - in: body
        name: request
        description: Request with the requirements to train the model
        required: true
        schema:
          $ref: '#/definitions/Requirements list'
      - name: url
        in: query
        description: The endpoint where the result of the operation will be returned
        required: false
        type: string
        allowEmptyValue: false
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResultId'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
      - binary-classification-controller
      summary: Update a model
      description: 'Given a list of requirements, updates the model of the classifier for the given company'
      operationId: updateUsingPUT
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - name: company
        in: query
        description: Proprietary company of the model
        required: true
        type: string
        allowEmptyValue: false
        x-example: UPC
      - name: property
        in: query
        description: Property of the classifier (i.e. property value of the *requirement_type* field)
        required: true
        type: string
        allowEmptyValue: false
        x-example: requirement
      - in: body
        name: request
        description: Request with the requirements to train and update the model
        required: true
        schema:
          $ref: '#/definitions/Requirements list'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
      - binary-classification-controller
      summary: Delete a model
      description: |-
        Given a **company** and a **property**, deletes the associated stored model. Additionally this method allows some variations:
        - If *property* = "ALL", removes all models of the given company.
        - If *company* = "ALL", removes all models of all companies (used as safe drop database).
      operationId: deleteModelUsingDELETE
      produces:
      - '*/*'
      parameters:
      - name: company
        in: query
        description: Proprietary company of the model
        required: true
        type: string
        allowEmptyValue: false
        x-example: UPC
      - name: property
        in: query
        description: Property of the classifier (i.e. property value of the *requirement_type* field)
        required: true
        type: string
        allowEmptyValue: false
        x-example: requirement
      responses:
        '200':
          description: Files deleted correctly
          schema:
            type: string
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Model(s) not found
          schema:
            type: string
      deprecated: false
  /upc/classifier-component/multiclassifier/classify:
    post:
      tags:
      - multilabel-classification-controller
      summary: Classify by domain
      description: 'Given a list of requirements, a company name and a domain of the company, classifies the list of requirements using the domain model. The result is a list of recommendations based on the classification results.'
      operationId: classifyUsingPOST_1
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - name: Model list
        in: query
        description: 'List of property values to generate models (if empty, all values are generated)'
        required: false
        type: array
        items:
          type: string
        collectionFormat: multi
        allowEmptyValue: false
      - name: company
        in: query
        description: Company to which the model belong
        required: true
        type: string
        allowEmptyValue: false
        x-example: UPC
      - name: context
        in: query
        description: 'Whether to apply contextual information analysis or not. If positive, contextual information (i.e. requirements'' position order and hierarchical structure) is used to apply the same tag during the classification process to all members belonging to a same document list structure.'
        required: false
        type: boolean
        default: false
        allowEmptyValue: false
        x-example: false
      - name: property
        in: query
        description: Property of the classifier
        required: true
        type: string
        allowEmptyValue: false
        x-example: requirement
      - in: body
        name: request
        description: Request with the requirements to train
        required: true
        schema:
          $ref: '#/definitions/Classify requirements list'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/RecommendationList'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /upc/classifier-component/multiclassifier/train&test:
    post:
      tags:
      - multilabel-classification-controller
      summary: Train and test by domain
      description: |
        Returns the result of k cross-validation using the requirements recieved in the request and the model of the implicit *company* and *property* (send as parameters in the request). Splits the requirements in k groups, trains a classifier for each group with all of the requirements recieved except the ones in the group and tests it with the requirements in the group.
        Returns the average of several statistics like the accuracy of the model
      operationId: trainAndTestUsingPOST_1
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: Model list
        in: query
        description: 'List of property values to generate models (if empty, all values are generated)'
        required: false
        type: array
        items:
          type: string
        collectionFormat: multi
        allowEmptyValue: false
      - name: context
        in: query
        description: 'Whether to apply contextual information analysis or not. If positive, contextual information (i.e. requirements'' position order and hierarchical structure) is used to apply the same tag during the classification process to all members belonging to a same document list structure.'
        required: false
        type: boolean
        default: false
        allowEmptyValue: false
        x-example: false
      - name: k
        in: query
        description: Number of tests
        required: true
        type: integer
        format: int32
        allowEmptyValue: false
        x-example: 10
      - name: property
        in: query
        description: Property of the classifier
        required: true
        type: string
        allowEmptyValue: false
        x-example: requirement
      - in: body
        name: request
        description: Request with the requirements to test
        required: true
        schema:
          $ref: '#/definitions/Multiclassify requirements list'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/DomainStats'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /upc/classifier-component/train&test:
    post:
      tags:
      - binary-classification-controller
      summary: Train and test
      description: |
        Returns the result of k cross-validation using the requirements recieved in the request. Splits the requirements in k groups, trains a classifier for each group with all of the requirements recieved except the ones in the group and tests it with the requirements in the group.
        Returns the average of several statistics like the accuracy of the model
      operationId: trainAndTestUsingPOST
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: context
        in: query
        description: 'Whether to apply contextual information analysis or not. If positive, contextual information (i.e. requirements'' position order and hierarchical structure) is used to apply the same tag during the classification process to all members belonging to a same document list structure.'
        required: false
        type: boolean
        default: false
        allowEmptyValue: false
        x-example: false
      - name: k
        in: query
        description: Number of tests
        required: true
        type: integer
        format: int32
        allowEmptyValue: false
      - name: property
        in: query
        description: Property of the classifier (i.e. property value of the *requirement_type* field)
        required: false
        type: string
        allowEmptyValue: false
      - in: body
        name: request
        description: Request with the requirements to test
        required: true
        schema:
          $ref: '#/definitions/Requirements list'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Stats'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
...
  ConfusionMatrixStats:
    type: object
    properties:
      false_negatives:
        type: integer
        format: int32
        description: False negatives
      false_positives:
        type: integer
        format: int32
        description: False positives
      true_negatives:
        type: integer
        format: int32
        description: True negatives
      true_positives:
        type: integer
        format: int32
        description: True positives
    title: ConfusionMatrixStats
  DomainStats:
    type: object
    properties:
      accuracy:
        type: number
        format: double
        description: Accuracy
      confusion_matrix:
        type: object
        description: Confusion matrix
        additionalProperties:
          $ref: '#/definitions/ConfusionMatrixStats'
      kappa:
        type: number
        format: double
        description: Kappa
      reliability:
        type: number
        format: double
        description: Reliability
      reliability_std_deviation:
        type: number
        format: double
        description: Reliability standard deviation
      weighted_f1_score:
        type: number
        format: double
        description: Weighted F1 score
      weighted_precision:
        type: number
        format: double
        description: Weighted precision
      weighted_recall:
        type: number
        format: double
        description: Weighted recall
    title: DomainStats
...
  Recommendation:
    type: object
    properties:
      confidence:
        type: number
        format: double
      requirement:
        type: string
      requirement_type:
        type: string
    title: Recommendation
  RecommendationList:
    type: object
    properties:
      recommendations:
        type: array
        items:
          $ref: '#/definitions/Recommendation'
    title: RecommendationList
...
  ResultId:
    type: object
    properties:
      id:
        type: string
    title: ResultId
  Stats:
    type: object
    properties:
      accuracy:
        type: number
        format: double
        description: Accuracy
      false_negatives:
        type: integer
        format: int32
        description: False negatives
      false_positives:
        type: integer
        format: int32
        description: False positives
      kappa:
        type: number
        format: double
        description: Kappa
      reliability:
        type: number
        format: double
        description: Reliability
      reliability_std_deviation:
        type: number
        format: double
        description: Reliability standard deviation
      true_negatives:
        type: integer
        format: int32
        description: True negatives
      true_positives:
        type: integer
        format: int32
        description: True positives
      weighted_f1_score:
        type: number
        format: double
        description: Weighted F1 score
      weighted_precision:
        type: number
        format: double
        description: Weighted precision
      weighted_recall:
        type: number
        format: double
        description: Weighted recall
    title: Stats
    description: Stats results of the classifier test